[clang-tidy] Renamed mesos-flags check to mesos-flags-inheritance.
