+++ clang-tidy/mesos/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(LLVM_LINK_COMPONENTS support)
 
 add_clang_library(clangTidyMesosModule
-  FlagsCheck.cpp
+  FlagsInheritanceCheck.cpp
   MesosTidyModule.cpp
   NamespaceCommentCheck.cpp
   ThisCaptureCheck.cpp
